# Prueba.WebServices/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Prueba.WebServices",
    "version": "1.0"
  },
  "paths": {
    "/Corresponsales/Index": {
      "get": {
        "tags": [
          "Corresponsales"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Corresponsales/Details/{id}": {
      "get": {
        "tags": [
          "Corresponsales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Corresponsales/Create": {
      "post": {
        "tags": [
          "Corresponsales"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Corresponsal"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Corresponsal"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Corresponsal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Corresponsales/Edit/{id}": {
      "post": {
        "tags": [
          "Corresponsales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Corresponsal"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Corresponsal"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Corresponsal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Corresponsales": {
      "post": {
        "tags": [
          "Corresponsales"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Oficinas/Index": {
      "get": {
        "tags": [
          "Oficinas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Oficinas/Details/{id}": {
      "get": {
        "tags": [
          "Oficinas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Oficinas/Create": {
      "post": {
        "tags": [
          "Oficinas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Oficina"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Oficina"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Oficina"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Oficinas/Edit/{id}": {
      "post": {
        "tags": [
          "Oficinas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Oficina"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Oficina"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Oficina"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Oficinas": {
      "post": {
        "tags": [
          "Oficinas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Corresponsal": {
        "type": "object",
        "properties": {
          "corCorresponsalId": {
            "type": "integer",
            "format": "int64"
          },
          "corNombre": {
            "type": "string",
            "nullable": true
          },
          "oficinas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Oficina"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Oficina": {
        "type": "object",
        "properties": {
          "ofiId": {
            "type": "integer",
            "format": "int64"
          },
          "ofiCorresponsalId": {
            "type": "integer",
            "format": "int64"
          },
          "ofiNombre": {
            "type": "string",
            "nullable": true
          },
          "ofiCorresponsal": {
            "$ref": "#/components/schemas/Corresponsal"
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}